Removed unused composeFiles from client.go
This PR removes the `composeFiles` variable from `compose()` and its use
in all the calling functions.

I liked how the CLI used a directory structure for instrumenting the
docker-compose files, and knew that Compose itself didn't support that,
so was curious what it was doing under the hood. In looking over the
code, I saw that the `composeFiles` variable was never used, so figured
I'd remove it and clean it up.