[FIX] Allow mnemonic recovery via file
The only way to `recover` was to specify the mnemonic as an input via
STDIN. This carries a number of risks, as that mnemonic could be exposed
via process monitoring, shell history, or over the shoulder observation.

I believe allowing the user to specify a file to recover from is a generally more
secure path to recovery from a mnemonic.